pyproject.toml
[project]
name = "streamlit-langgraph-toolcall"
version = "0.1.0"
description = "A simple chatbot showcasing integration to langraph and tool calling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.7",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
